Add an orange background to the toggle when it's active
This helps convey the current state.

With the toggle, you can choose to use words to label one of two things:

1. What will happen when you activate the toggle i.e. the Action
2. What is currently happening i.e. the Status

I think labelling the action is better, specifically if you think about screen readers. Those treat items like toggles the same as buttons; buttons should be labeled with verbs.

However, we still want to convey the current status somehow. Here I've chosen to convey the status both with a hidden tag (previous commit) which allows screen readers to find the status and to label the "active" status with a bright color in the background of the toggle.