Implement countdown timer
This shifts the experience from counting up to counting down.

The default countdown is 5minutes (which should probably be configurable in future).

The timer can be edited when the timer is off. It is currently only possible to edit the timer in seconds. Minutes, hours and days are very much expected to be added later.